command groups, lambda logging, lambda deployment & invokcation both local & in cloud
