+++ cdk/requirements/base.txt
@@ -214,7 +214,9 @@ aws-cdk.region-info==1.95.1
     #   aws-cdk.aws-iam
     #   aws-cdk.core
 aws-lambda-logging==0.1.1
-    # via -r requirements/base.in
+    # via
+    #   -r requirements/../image/requirements.in
+    #   -r requirements/base.in
 boto3==1.17.39
     # via -r requirements/base.in
 botocore==1.20.39
@@ -223,6 +225,10 @@ botocore==1.20.39
     #   s3transfer
 cattrs==1.3.0
     # via jsii
+certifi==2020.12.5
+    # via requests
+chardet==4.0.0
+    # via requests
 click==7.1.2
     # via -r requirements/base.in
 constructs==3.3.71
@@ -256,6 +262,8 @@ constructs==3.3.71
     #   aws-cdk.aws-ssm
     #   aws-cdk.core
     #   aws-cdk.custom-resources
+idna==2.10
+    # via requests
 jmespath==0.10.0
     # via
     #   boto3
@@ -336,6 +344,8 @@ python-dateutil==2.8.1
     #   jsii
 python-dotenv==0.16.0
     # via -r requirements/base.in
+requests==2.25.1
+    # via -r requirements/base.in
 s3transfer==0.3.6
     # via boto3
 six==1.15.0
@@ -343,4 +353,6 @@ six==1.15.0
 typing-extensions==3.7.4.3
     # via jsii
 urllib3==1.26.4
-    # via botocore
+    # via
+    #   botocore
+    #   requests
