+++ cdk/requirements/dev.txt
@@ -220,7 +220,9 @@ aws-cdk.region-info==1.95.1
     #   aws-cdk.aws-iam
     #   aws-cdk.core
 aws-lambda-logging==0.1.1
-    # via -r requirements/base.in
+    # via
+    #   -r requirements/../image/requirements.in
+    #   -r requirements/base.in
 black==20.8b1
     # via -r requirements/dev.in
 boto3==1.17.39
@@ -231,8 +233,12 @@ botocore==1.20.39
     #   s3transfer
 cattrs==1.3.0
     # via jsii
+certifi==2020.12.5
+    # via requests
 cfgv==3.2.0
     # via pre-commit
+chardet==4.0.0
+    # via requests
 click==7.1.2
     # via
     #   -r requirements/base.in
@@ -277,6 +283,8 @@ flake8==3.9.0
     # via -r requirements/dev.in
 identify==2.2.1
     # via pre-commit
+idna==2.10
+    # via requests
 jmespath==0.10.0
     # via
     #   boto3
@@ -375,6 +383,8 @@ pyyaml==5.4.1
     # via pre-commit
 regex==2021.3.17
     # via black
+requests==2.25.1
+    # via -r requirements/base.in
 s3transfer==0.3.6
     # via boto3
 six==1.15.0
@@ -392,6 +402,8 @@ typing-extensions==3.7.4.3
     #   black
     #   jsii
 urllib3==1.26.4
-    # via botocore
+    # via
+    #   botocore
+    #   requests
 virtualenv==20.4.3
     # via pre-commit
